Replace Unix.fork and Unix.wait calls for Windows.
Unix.fork and Unix.wait are not available on Windows. OpamParallel updated
to invoke a new process and pass the child closure using Marshal instead.
Unix.waitpid is not fully implemented on Windows, sadly (being able to
pass -1 would be sufficient) so a C stub is added to wait specifically on
the list of pids.