Windows manipulation of parent process environment.
The Windows Command prompt doesn't have an obvious equivalent to sh's
eval. Two Windows-specific functions added to OpamMisc. parent_putenv is
the same as Unix.putenv, except that it manipulates the parent-process's
environment (which, for opam.exe, will be the Command Prompt).
persistHomeDirectory allows $HOME to be permanently set to a given value
for the current user (and also updates the parent process's environment).