Windows-on-Windows support.
In 64-bit Windows, a injecting 32-bit code into a 64-bit process will
silently fail and injecting 64-bit code into a 32-bit process will
kill the process. It is relatively easy for a 64-bit process to inject
32-bit code into its parent but it's incredibly hard to do the reverse and
both mechanisms are fairly brittle (the curious reader is invited to
search for Heaven's Gate).

This patch adds the preliminary work for opam-putenv (installed under
Windows only). For a 64-bit build of OPAM, this tool is compiled with a
32-bit C compiler and vice versa (configure machinery not yet added). This
process is passed the PID and a set of environment variable names and
values on the command line and does the process injection itself. It is
used if OPAM detects an architecture mismatch between itself and its
parent.