Clean up broken conflation of URIs and Filenames.
OPAM treats URIs and Filenames as though they're the same object. While
this is quite neat for the repository backend and of course works fine on
Unix where the separators are the same, this completely breaks on Windows
with the backslash default separator.

This isn't the greatest attempt to deal with this - to minimise the
changes, a noisy incompatible type has been introduced to the download
functions with various coercion functions. Finally, OpamFilename.t now
includes the separator in the record, replacing most calls to
Filename.concat with its own implementation.