Fix logging problem on Windows.
Unix.getpid doesn't return the PID on Windows (see PR 4034) which means
it's not useful for creating a logging directory once using OpamParallel!
Fixed by creating an additional C stub to get the actual PID. This may get
fixed in OCaml's Unix at some point but, as the discussion on the PR
notes, fixing Unix.getpid is easy but maintaining the Unix semantics of a
PID on Windows is much harder.