Second attempt to generalise OpamFilename.
Add an extra parameter to OpamFilename.Dir.to_string and
OpamFilename.to_string to ensure that in all call locations the nature of
the use of the filename is considered. Intentionally caused all code to
need updating, rather than using an optional argument. Some further
refinement needed (in particular combining with the previous attempt may
be useful for rsync remotes and URIs -- i.e. be able to create
OpamFilename.t objects which use a forward slash in all profile contexts).