Add system support to switch label for Windows.
Permit an extra +tag at the end of a switch name to indicate which Windows
port is installed and report it in the output of opam switch list.

The four ports are named msvc, msvc64, mingw and mingw64 so if all four
ports of 4.02.1 were installed, switches 4.02.1+msvc, 4.02.1+msvc64,
4.02.1+mingw and 4.02.1+mingw64 would be created.

The parsing is intended to work along side PR switches (e.g.
4.02.1+pr666+msvc).