Lock-free 'opam config env' and 'opam config exec'
On 2-Dec-2014 @ 9:14:14 Louis Gesbert wrote:
> Since it's more likely to cause than to avoid problems, let's be pragmatic.
> Closes #1830

The problem with 39c9ee70bd210a54f85f722721d37156c3c56d98 is that it
skips the check at the start of OpamState.check that ~/.opam exists. As
a result, if ~/.opam doesn't exist, an invalid directory is created in
its place.

Introduced the Without_lock lock, so that OpamState.check can still be
called for opam config env and opam config exec.