Set and persist $HOME on Windows in opam init
Windows setups will not normally have $HOME set, which causes opam init
to initialise in the cwd. Native Windows updated to query the user's
Documents folder and use that location instead.

Additionally, opam init offers to save $HOME permanently to the user's
registry, as having HOME set is important for OCaml (so that .ocamlinit
can be found) and also for related tools (Vim, for example, can display
and use ~\ paths if $HOME is set)