Set Windows Command Processor AutoRun entry
Allow opam init to set the user's registry so that the Windows Command
Processor automatically executes opam config env every time a command
prompt is launched.

This setting has potential to be very problematic (which is why the
change is in a separate commit from the main work supporting the Windows
Command Prompt). It is absolutely imperative that opam produces no
output to stdout or stderr when executing from the AutoRun key since the
AutoRun entry is executed even for sub-commands (cmd /c or, importantly,
Sys.command!) and for this reason a hidden flag with special parsing is
added so that opam config env --autorun is what is actually added to the
registry.

If the user uninstalls OPAM or removes opam.exe from their PATH, it is
imperative that this registry entry is removed.